sig: simple config (#75)
* sig: remove default prometheus address (#3850)

The SIG was setting a default prometheus address, contrary to what is
documented; the docs explicitly say that if the address is unset, the
metrics will not be exported. This default value would also make it
impossible to disable exporting metrics.

Remove the undocumented, unneeded default value.

* sig: remove IA from config, make IP optional (#3849)

The local IA in the sig.toml config is redundant as it can be obtained
from sciond. Simply removed the option.
The bind IP option is made optional by defaulting to a sensible local
IP, using the same logic used e.g. in showpaths.

Both of these changes combined allow us to package a default sig.toml
config file that (in most cases) "just works" without modifications.

* sig: keep isd_as option for backwards compatibility